+++ folio-secret-store/folio-secret-store-common/src/test/java/org/folio/tools/store/impl/EphemeralStoreTest.java
@@ -8,7 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.Map;
 import java.util.Properties;
 import org.folio.test.types.UnitTest;
-import org.folio.tools.store.exception.NotFoundException;
+import org.folio.tools.store.exception.SecretNotFoundException;
 import org.folio.tools.store.properties.EphemeralConfigProperties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -39,7 +39,7 @@ class EphemeralStoreTest {
   void get_negative_notFound() {
     var key = "baz";
 
-    assertThrows(NotFoundException.class, () -> ephemeralStore.get(key));
+    assertThrows(SecretNotFoundException.class, () -> ephemeralStore.get(key));
   }
 
   @Test
@@ -73,13 +73,25 @@ class EphemeralStoreTest {
   }
 
   @Test
+  void set_positive_nullValueDeletesKey() {
+    var key = "foo";
+    var value = "bar";
+    ephemeralStore.set(key, value);
+    assertEquals(value, ephemeralStore.get(key));
+
+    ephemeralStore.set(key, null);
+    // After setting null, the key should be deleted
+    assertThrows(SecretNotFoundException.class, () -> ephemeralStore.get(key));
+  }
+
+  @Test
   void set_negative() {
     var key = "foo";
     var value = "bar";
 
     ephemeralStore.set(key, value);
 
-    assertThrows(NotFoundException.class, () -> ephemeralStore.get("baz"));
+    assertThrows(SecretNotFoundException.class, () -> ephemeralStore.get("baz"));
   }
 
   @Test
@@ -93,10 +105,10 @@ class EphemeralStoreTest {
 
     var store = new EphemeralStore(props);
 
-    assertEquals("dit_password", store.get(null, "dit", "dit"));
-    assertEquals("dot_password", store.get(null, "dot", "dot"));
-    assertEquals("dat_password", store.get(null, "dat", "dat"));
-    assertEquals("", store.get(null, "done", "done"));
+    assertEquals("dit_password", store.get("dit_dit"));
+    assertEquals("dot_password", store.get("dot_dot"));
+    assertEquals("dat_password", store.get("dat_dat"));
+    assertEquals("", store.get("done_done"));
   }
 
   @Test
@@ -110,9 +122,35 @@ class EphemeralStoreTest {
 
     var store = new EphemeralStore(props);
 
-    assertThrows(NotFoundException.class, () -> store.get(null, "dit", "foo"));
-    assertThrows(NotFoundException.class, () -> store.get(null, "dot", "foo"));
-    assertThrows(NotFoundException.class, () -> store.get(null, "dat", "foo"));
-    assertThrows(NotFoundException.class, () -> store.get(null, "done", "foo"));
+    assertThrows(SecretNotFoundException.class, () -> store.get("dit_foo"));
+    assertThrows(SecretNotFoundException.class, () -> store.get("dot_foo"));
+    assertThrows(SecretNotFoundException.class, () -> store.get("dat_foo"));
+    assertThrows(SecretNotFoundException.class, () -> store.get("done_foo"));
+  }
+
+  @Test
+  void delete_positive() {
+    var key = "foo";
+    var value = "bar";
+    ephemeralStore.set(key, value);
+
+    ephemeralStore.delete(key);
+
+    assertThrows(SecretNotFoundException.class, () -> ephemeralStore.get(key));
+  }
+
+  @Test
+  void delete_negative_keyNotPresent() {
+    var key = "not_existing_key";
+    // Should not throw any exception
+    ephemeralStore.delete(key);
+    // Still not present
+    assertThrows(SecretNotFoundException.class, () -> ephemeralStore.get(key));
+  }
+
+  @Test
+  void delete_negative_invalidKey() {
+    assertThrows(IllegalArgumentException.class, () -> ephemeralStore.delete(""));
+    assertThrows(IllegalArgumentException.class, () -> ephemeralStore.delete(null));
   }
 }